Add files via upload
Progress : 
   - CRUD user selesai
   - Jadwal Dokter, Progress
   - Tagihan, Progress

+++ application/controllers/Admin.php
@@ -14,7 +14,7 @@
 
 		// function input data
 
-		function input_data_dokter(){
+		function input_data_dokter($mod = 0){
 			$this->form_validation->set_rules('txtId', 'ID','required');
 			$this->form_validation->set_rules('txtNama', 'Nama','required', array('required' => 'Nama Tidak Boleh Kosong'));
 			$this->form_validation->set_rules('txtNoTelp', 'No. Telpon','required', array('required' => 'No. Telpon Tidak Boleh Kosong'));
@@ -42,11 +42,20 @@
 						'ID_POLIKLINIK' => $this->input->post('txtId')
 					)
 				);
-				$dokter = $this->admin_model->insert_dokter($data['data_dokter']);
-				$user = $this->admin_model->insert_user($data['data_user']);
-				if($dokter == TRUE && $user == TRUE){
-					$this->show_input_dokter("Data Ditambahkan");
+				if($mod != 0){
+					$dokter = $this->admin_model->update_dokter($data['data_dokter']);
+					$user = $this->admin_model->update_user($data['data_user']);
+					if($dokter == TRUE || $user == TRUE){
+						$this->show_input_dokter("Data Berhasil Dirubah");
+					}
+				}else{
+					$dokter = $this->admin_model->insert_dokter($data['data_dokter']);
+					$user = $this->admin_model->insert_user($data['data_user']);
+					if($dokter == TRUE || $user == TRUE){
+						$this->show_input_dokter("Data Berhasil Dirubah");
+					}	
 				}
+				
 			}
 		}
 
@@ -77,9 +86,9 @@
 						'ID_POLIKLINIK' => $this->input->post('txtId')
 					)
 				);
-				$dokter = $this->admin_model->insert_perawat($data['data_perawat']);
+				$perawat = $this->admin_model->insert_perawat($data['data_perawat']);
 				$user = $this->admin_model->insert_user($data['data_user']);
-				if($dokter == TRUE && $user == TRUE){
+				if($perawat == TRUE && $user == TRUE){
 					$this->show_input_perawat("Data Ditambahkan");
 				}
 			}
@@ -87,6 +96,21 @@
 
 		function input_data_poliklinik(){
 			$this->form_validation->set_rules('txtIdPoli','ID Poliklinik', 'required');
+			$this->form_validation->set_rules('txtNamaPoli','Nama Poliklinik', 'required',array('required' => 'Nama Poliklinik tidak boleh kosong'));
+			if($this->form_validation->run() == FALSE){
+				$this->show_input_poliklinik();
+				return;
+			}
+			$data = array(
+				'ID_POLI' => $this->input->post('txtIdPoli'),
+				'NAMA_POLI' => $this->input->post('txtNamaPoli')
+			);
+			$res = $this->admin_model->insert_poli($data);
+			$this->show_input_poliklinik($res);
+		}
+
+		function input_data_jam(){
+			$this->form_validation->set_rules('txtIdJam','Id Jam', 'required');
 			$this->form_validation->set_rules('txtNamaPoli','Nama Poliklinik', 'required',array('required' => 'Nama Poliklinik tidak boleh kosong'));
 			if($this->form_validation->run() == FALSE){
 				$this->show_input_poliklinik();
@@ -128,9 +152,75 @@
 			$this->show_view_poliklinik();
 		}
 
+		function delete_data_dokter($ID_DOKTER){
+			$res1 = $this->admin_model->delete_dokter($ID_DOKTER);
+			$res2 = $this->admin_model->delete_user($ID_DOKTER);
+			if($res1 == TRUE && $res2 == TRUE){
+				$this->show_view_user("Data Berhasil Dihapus");	
+			}
+		}
+
+		function delete_data_perawat($ID_PERAWAT){
+			$res = $this->admin_model->delete_perawat($ID_PERAWAT);
+			$res2 = $this->admin_model->delete_user($ID_PERAWAT);
+			if($res1 == TRUE && $res2 == TRUE){
+				$this->show_view_user("Data Berhasil Dihapus");	
+			}
+		}
+
 		// end function delete
 
 		//function get_data
+
+		function get_data_dokter(){
+			$list = $this->admin_model->get_list_dokter();
+			echo "<table>
+					<tr>
+						<th>ID DOKTER</td>
+						<th>NAMA</td>
+						<th>ALAMAT</td>
+						<th>NO. TELP</td>
+						<th>ID POLI</td>
+						<th>ACTION</td>
+					</tr>";
+			foreach($list as $row){
+				echo "<tr>
+						<td>".$row['ID_DOKTER']."</td>
+						<td>".$row['NAMA']."</td>
+						<td>".$row['ALAMAT']."</td>
+						<td>".$row['NO_TELP']."</td>
+						<td>".$row['NAMA_POLI']."</td>";
+					echo "<td><a href=\"".base_url()."admin/show_jadwal_dokter/".$row['ID_DOKTER']."\">JADWAL</a>|<a href=\"".base_url()."admin/show_update_dokter/".$row['ID_DOKTER']."\">EDIT</a>|<a href=\"".base_url()."admin/delete_data_/".$row['ID_DOKTER']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>";
+				echo "</tr>";
+			}
+			echo "</table>";
+		}
+
+		function get_single_dokter($data){
+			$list = $this->admin_model->get_specific_dokter($data);
+			echo "<table>
+					<tr>
+						<th>ID DOKTER</td>
+						<th>NAMA</td>
+						<th>ALAMAT</td>
+						<th>NO. TELP</td>
+						<th>ID POLI</td>
+						<th>ACTION</td>
+					</tr>";
+			if($data == "kosong"){
+				echo "kosong";
+			}
+			foreach($list as $row){
+				echo "<tr>
+						<td>".$row['ID_DOKTER']."</td>
+						<td>".$row['NAMA']."</td>
+						<td>".$row['ALAMAT']."</td>
+						<td>".$row['NO_TELP']."</td>
+						<td>".$row['NAMA_POLI']."</td>";
+					echo "<td><a href=\"".base_url()."admin/show_jadwal_dokter/".$row['ID_DOKTER']."\">JADWAL</a>|<a href=\"".base_url()."admin/show_update_dokter/".$row['ID_DOKTER']."\">EDIT</a>|<a href=\"".base_url()."admin/delete_data_/".$row['ID_DOKTER']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>";
+			}
+			echo "</table>";
+		}
 
 		function get_data_user(){
 			$list = $this->admin_model->get_list_user();
@@ -151,7 +241,7 @@
 				if($row['TIPE'] == "admin"){
 					echo "<td>Tidak Tersedia</td>";
 				}else{
-					echo "<td><a href=\"".base_url()."admin/show_update_".$row['TIPE']."/".$row['ID_USER']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_".$row['TIPE']."/".$row['ID_USER']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>";
+					echo "<td><a href=\"".base_url()."admin/show_update_".$row['TIPE']."/".$row['ID_POLIKLINIK']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_".$row['TIPE']."/".$row['ID_POLIKLINIK']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>";
 				}
 				echo "</tr>";
 			}
@@ -174,9 +264,13 @@
 							<td>".$row['ID_USER']."</td>
 							<td>".$row['ID_POLIKLINIK']."</td>
 							<td>".$row['USERNAME']."</td>
-							<td>".$row['TIPE']."</td>
-							<td><a href=\"".base_url()."admin/show_update_".$row['TIPE']."/".$row['ID_USER']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_".$row['TIPE']."/".$row['ID_USER']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>			
-						</tr>";
+							<td>".$row['TIPE']."</td>";
+					if($row['TIPE'] == "admin"){
+							echo "<td>Tidak Tersedia</td>";
+						}else{
+							echo "<td><a href=\"".base_url()."admin/show_update_".$row['TIPE']."/".$row['ID_POLIKLINIK']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_".$row['TIPE']."/".$row['ID_POLIKLINIK']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>";
+						}
+						echo "</tr>";
 				}
 				echo "</table>";
 			}else{
@@ -197,6 +291,33 @@
 					echo "<tr>
 							<td>".$row['ID_POLI']."</td>
 							<td>".$row['NAMA_POLI']."</td>
+							<td><a href=\"".base_url()."admin/show_update_poliklinik/".$row['ID_POLI']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_poliklinik/".$row['ID_POLI']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>								
+						</tr>";
+				}
+				echo "</table>";
+			}else{
+				echo "<table>
+					<tr>
+						<td>Data Kosong</td>
+					</tr>
+					";
+			}
+
+		}
+
+		function get_data_jam(){
+			$list = $this->admin_model->get_list_jam();
+			if($list != "kosong"){
+				echo "<table>
+						<tr>
+							<th>ID JAM</td>
+							<th>JAM</td>
+							<th>Action</td>
+						</tr>";
+				foreach($list as $row){
+					echo "<tr>
+							<td>".$row['ID_JAM']."</td>
+							<td>".$row['JAM']."</td>
 							<td><a href=\"".base_url()."admin/show_update_poliklinik/".$row['ID_POLI']."\">EDIT</a> | <a href=\"".base_url()."admin/delete_data_poliklinik/".$row['ID_POLI']."\" onclick=\"return confirm('Are you sure?')\">DELETE</a></td>								
 						</tr>";
 				}
@@ -253,10 +374,24 @@
 			$this->load->view('template/footer');
 		}
 
-		function show_view_user(){
+		function show_input_jadwal(){
+			$data['kode'] = $this->admin_model->generate_id_jadwal();
+			echO $data['kode'];
 			$this->load->view('template/header');
 			$this->load->view('admin/nav_admin');
-			$this->load->view('admin/edituser');
+			$this->load->view('admin/input_jadwal', $data);
+			$this->load->view('template/footer');
+		}
+
+		function show_view_user($msg = ""){
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			if($msg != ""){
+				$data['msg'] = $msg;
+				$this->load->view('admin/edituser', $data);
+			}else{
+				$this->load->view('admin/edituser');
+			}
 			$this->load->view('template/footer');
 		}
 
@@ -271,10 +406,35 @@
 			$this->load->view('template/footer');
 		}
 
+		function show_input_jam($res = FALSE){
+			$data['kode'] = $this->admin_model->generate_id_jam();
+			if($res == TRUE){
+				$data['res'] = $res;
+			}
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			$this->load->view('admin/input_jam', $data);
+			$this->load->view('template/footer');
+		}
+
 		function show_view_poliklinik(){
 			$this->load->view('template/header');
 			$this->load->view('admin/nav_admin');
 			$this->load->view('admin/editpoliklinik');
+			$this->load->view('template/footer');
+		}
+
+		function show_view_jam(){
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			$this->load->view('admin/editjam');
+			$this->load->view('template/footer');
+		}
+
+		function show_view_dokter(){
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			$this->load->view('admin/editdokter');
 			$this->load->view('template/footer');
 		}
 
@@ -290,17 +450,47 @@
 		}
 
 		function show_update_dokter($kode, $res= FALSE){
-			$data_user= $this->admin_model->get_specific_user($kode);
-			$data_dokter = $this->admin_model->get_specific_dokter($data_user['ID_POLIKLINIK']);
-			echo $data_dokter['NAMA'];
-			$data['data_user'] = $data_user;
+			// $data_user= $this->admin_model->get_specific_user($kode);
+			$data_dokter = $this->admin_model->get_specific_dokter($kode);
+			$data_poliklinik = $this->admin_model->get_list_poli();
+			//$data['data_user'] = $data_user;
 			$data['data_dokter'] = $data_dokter;
+			$data['data_poliklinik'] = $data_poliklinik;
 			if($res == TRUE){
 				$data['res'] = $res;
 			}
 			$this->load->view('template/header');
 			$this->load->view('admin/nav_admin');
 			$this->load->view('admin/update_data_dokter', $data);
+			$this->load->view('template/footer');
+		}
+		function show_update_perawat($kode, $res= FALSE){
+			$data_user= $this->admin_model->get_specific_user($kode);
+			$data_perawat = $this->admin_model->get_specific_perawat($data_user[0]['ID_POLIKLINIK']);
+			echo $data_user[0]['ID_POLIKLINIK'];
+			$data['data_user'] = $data_user;
+			$data['data_perawat'] = $data_perawat;
+			if($res == TRUE){
+				$data['res'] = $res;
+			}
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			$this->load->view('admin/update_data_perawat', $data);
+			$this->load->view('template/footer');
+		}
+
+		function show_jadwal_dokter($kode = ""){
+			$this->load->view('template/header');
+			$this->load->view('admin/nav_admin');
+			if($kode == ""){
+				$this->load->view('admin/data_dokter', $data['msg'] = "kosong");
+			}else{
+				$data_dokter = $this->admin_model->get_specific_dokter($kode);
+				$data_jadwal = $this->admin_model->get_list_jadwal($kode);
+				$data['data_dokter'] = $data_dokter;
+				$data['data_jadwal'] = $data_jadwal;
+				$this->load->view('admin/data_dokter', $data);
+			}
 			$this->load->view('template/footer');
 		}
 		// end show function
